feature: refactoring develop branch in order to integrate another connection.
