+++ src/Services/Present/ShopService.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Epayco\SdkRedeban\Services;
+namespace Epayco\SdkRedeban\Services\Present;
 
 use Epayco\SdkRedeban\Repositories\RedebanRepository;
+use Epayco\SdkRedeban\Services\Service;
 use Exception;
 use stdClass;
