+++ src/Services/Service.php
@@ -1,4 +1,5 @@
-<?php 
+<?php
+
 namespace Epayco\SdkRedeban\Services;
 class Service{
